1.support RESOURCE Create\Delete\Update\Query for Mysql 2.support RESOURCE restFul(part)

+++ auth/src/main/java/com/rcircle/service/auth/model/Account.java
@@ -6,7 +6,7 @@ import java.io.Serializable;
 import java.util.List;
 
 public class Account implements Serializable {
-    private long uid = 0;
+    private int uid = 0;
     private String username;
     private String password;
     private int status;
@@ -15,11 +15,11 @@ public class Account implements Serializable {
     private long lastlogin;
     private List<Authority> roles;
 
-    public long getUid() {
+    public int getUid() {
         return uid;
     }
 
-    public void setUid(long uid) {
+    public void setUid(int uid) {
         this.uid = uid;
     }
